Update ReadMe.md
I decided to fork WebKit after a long time of desperation, read the license, have fun, and read and understand the code. The software product that may come out of this shall not be relied at the moment. I have since decided to take a break and start painting instead.